Factor out token ranking
