--- rust-bors.toml
@@ -1,1 +1,0 @@
-timeout = 3600
